requirements.txt
blinker==1.6.3
bpython==0.24
brotli==1.1.0
brotlicffi==1.1.0.0
cffi==1.12.3
cffi==1.15.1
ConfigParser==6.0.0
cx_Oracle==8.3.0
Cython==0.29.12
dl==0.1.0
docutils==0.20.1
HTMLParser==0.0.2
httplib2==0.22.0
ipaddr==2.2.0
ipdb==0.13.13
ipython==7.6.1
mock==3.0.5
MySQL-python==1.2.5
numpy==1.16.4
onelogin==3.1.6
ordereddict==1.1
Pillow==6.1.0
Pillow==10.1.0
psycopg2==2.9.9
psycopg2cffi==2.9.0
pycryptodome==3.19.0
pycurl==7.45.2
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==19.0.0
pyOpenSSL==23.3.0
pytest==5.0.1
pytest_django==4.5.2
python_bcrypt==0.3.2
python_memcached==1.59
pywatchman==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
sets==0.3.2
simplejson==3.19.2
tblib==1.4.0
urllib3_secure_extra==0.1.0
wincertstore==0.2.1
xmlrpclib==1.0.1
zstandard==0.21.0
